+++ main/src/main/java/net/oneandone/stool/docker/Engine.java
@@ -8,20 +8,15 @@ import com.google.gson.JsonParser;
 import com.google.gson.JsonPrimitive;
 import jnr.unixsocket.UnixSocketAddress;
 import jnr.unixsocket.UnixSocketChannel;
-import net.oneandone.sushi.fs.FileNotFoundException;
 import net.oneandone.sushi.fs.Node;
 import net.oneandone.sushi.fs.World;
 import net.oneandone.sushi.fs.file.FileNode;
 import net.oneandone.sushi.fs.http.HttpFilesystem;
 import net.oneandone.sushi.fs.http.HttpNode;
-import net.oneandone.sushi.fs.http.MovedTemporarilyException;
 import net.oneandone.sushi.fs.http.StatusException;
 import net.oneandone.sushi.fs.http.io.AsciiInputStream;
 import net.oneandone.sushi.fs.http.model.Body;
 import net.oneandone.sushi.fs.http.model.Method;
-import net.oneandone.sushi.fs.http.model.Request;
-import net.oneandone.sushi.fs.http.model.Response;
-import net.oneandone.sushi.fs.http.model.StatusCode;
 import net.oneandone.sushi.util.Strings;
 import org.kamranzafar.jtar.TarEntry;
 import org.kamranzafar.jtar.TarHeader;
@@ -33,7 +28,6 @@ import java.io.ByteArrayOutputStream;
 import java.io.DataInputStream;
 import java.io.EOFException;
 import java.io.File;
-import java.io.FilterInputStream;
 import java.io.IOException;
 import java.io.InputStream;
 import java.io.OutputStream;
@@ -517,45 +511,8 @@ public class Engine {
     }
 
     private InputStream postStream(HttpNode dest, byte[] body) throws IOException {
-        return postStream(dest, new Body(null, null, body.length, new ByteArrayInputStream(body), false));
+        return dest.postStream(new Body(null, null, body.length, new ByteArrayInputStream(body), false));
     }
-
-    // TODO: move to sushi
-    private static InputStream postStream(HttpNode resource, Body body) throws IOException {
-        Request post;
-        Response response;
-
-        post = new Request("POST", resource);
-        post.bodyHeader(body);
-        response = post.responseHeader(post.open(body));
-        if (response.getStatusLine().code == StatusCode.OK) {
-            return new FilterInputStream(response.getBody().content) {
-                private boolean freed = false;
-
-                @Override
-                public void close() throws IOException {
-                    if (!freed) {
-                        freed = true;
-                        post.free(response);
-                    }
-                    super.close();
-                }
-            };
-        } else {
-            post.free(response);
-            switch (response.getStatusLine().code) {
-                case StatusCode.MOVED_TEMPORARILY:
-                    throw new MovedTemporarilyException(response.getHeaderList().getFirstValue("Location"));
-                case StatusCode.NOT_FOUND:
-                case StatusCode.GONE:
-                case StatusCode.MOVED_PERMANENTLY:
-                    throw new FileNotFoundException(resource);
-                default:
-                    throw StatusException.forResponse(response);
-            }
-        }
-    }
-
 
     private String post(HttpNode dest, String body) throws IOException {
         try {